Update intro-react.md
Edited the JSX paragraph for easy readability, changed "Because JSX is JavaScript, you can use variables inside it. Here you are declaring a name for the cat, name, and embedding it with curly braces inside <Text>." to "Since JSX is based on JavaScript, you can easily include variables in it. In the following code snippet, you're declaring a name for the cat, `name`, and embedding it within `<Text>` using curly braces." and also gave it proper spacing.